Update upgrade_all.pl
Adds ability to skip Space Stations based on their name using an special switch and string such as --skipSS "^(S|Z)ASS" to quickly eliminate those from the list before using up RPCs checking to see if they are stations. Couple other readability tweaks. When the --skipSS switch is used, those will not show up on the skipped planets printout as we already know we do not wish to look at them.  Adds extra newline char after clearing the queue printout so it is easier to differentiate when an new loop starts.